move to correct folder & create routes for pages
