+++ client/src/components/NavBar.js
@@ -7,10 +7,10 @@ export default function NavBar() {
     return (
         <Navbar bg="dark" variant="dark">
         <Container>
-          <Navbar.Brand href="#home">Climate Data</Navbar.Brand>
+          <Navbar.Brand href="/">Climate Data</Navbar.Brand>
           <Nav className="me-auto">
             <Nav.Link href="/">Temperature & CO2 Concentrations</Nav.Link>
-            <Nav.Link href="#features">Emissions</Nav.Link>
+            <Nav.Link href="/emissions">Emissions</Nav.Link>
             <Nav.Link href="#pricing">Custom Page</Nav.Link>
           </Nav>
         </Container>
