+++ client/src/components/RenderActiveShape.js
