+++ client/src/components/StackedLineChart.js
