+++ client/src/pages/Emissions.js
@@ -1,0 +1,10 @@
+import EmissionPieChart from "../components/EmissionPieChart";
+
+export default function Emissions(){
+
+    return(
+        <div>
+            <EmissionPieChart/>
+        </div>
+    )
+}
